Renaming of classes
